refactor: add new "Topics" component
