--- packages/gatsby-theme-project-portal/src/components/Topic.tsx
@@ -1,13 +1,0 @@
-import { graphql } from "gatsby"
-
-export interface TopicType {
-  slug: string
-  title: string
-}
-
-export const query = graphql`
-  fragment TopicDetails on Topic {
-    slug
-    title
-  }
-`
